requirements.txt
aiohttp
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
discord-py-interactions==4.4.1
discord-py-slash-command==3.0.1
discord.py
dislash.py==1.4.9
frozenlist==1.4.0
google-api-core==2.1.1
google-api-python-client==2.27.0
google-auth==2.3.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.53.0
gspread
httplib2==0.20.1
idna==3.2
multidict==5.1.0
numpy
oauthlib==3.1.1
protobuf==3.18.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyNaCl==1.4.0
pyparsing==2.4.7
python-dotenv==0.19.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
typing-extensions==3.10.0.0
uritemplate==4.1.1
urllib3==1.26.7
yarl
PyYAML